Refactor packages into smaller structures.
Related to #24
