+++ colmeia-bins/Cargo.toml
@@ -1,16 +1,13 @@
 [package]
-name = 'colmeia-hypercore'
+name = 'colmeia-bins'
 version = '0.1.0'
 authors = ['Bruno Tavares <connect+github@bltavares.com>']
 edition = '2018'
 
 [dependencies]
-futures = '0.3.5'
-hypercore = '0.11.1-beta.9'
-anyhow = '1.0.34'
-random-access-storage = '4.0.0'
-log = '0.4.8'
-bitfield-rle = '0.2.0'
+env_logger = '*'
+log = '*'
+hex = '*'
 
 [dependencies.async-std]
 version = '1.6.2'
@@ -18,3 +15,9 @@ features = ['unstable']
 
 [dependencies.hypercore-protocol]
 git = 'https://github.com/Frando/hypercore-protocol-rs'
+
+[dependencies.colmeia-hyperswarm-mdns]
+path = '../colmeia-hyperswarm-mdns'
+
+[dependencies.colmeia-hyperstack]
+path = '../colmeia-hyperstack'
