+++ colmeia-hypercore/README.md
@@ -1,0 +1,3 @@
+# colmeia-hypercore
+
+Network layer for [datrs/hypercore](https://github.com/datrs/hypercore).
