+++ colmeia-hypercore/src/lib.rs
@@ -1,10 +1,3 @@
-mod hyperdrive;
-mod hyperstack;
 mod network;
-mod schema;
-mod utils;
 
-pub use hyperdrive::{in_memmory, Hyperdrive};
-pub use hyperstack::Hyperstack;
-pub use network::{replicate_hyperdrive, Emit, PeeredFeed};
-pub use utils::{HashParserError, PublicKeyExt};
+pub use network::{Emit, PeeredFeed};
