+++ colmeia-hyperdrive/README.md
@@ -1,0 +1,5 @@
+# colmeia-hyperdrive
+
+Implementation of [hyper-protocol/hyperpdrive](https://github.com/hypercore-protocol/hyperdrive/) in Rust.
+
+Uses peered [colmeia-hypercore](../colmeia-hypercore) to sync the peered feeds: metadata and content.
