+++ colmeia-hyperdrive/src/lib.rs
@@ -1,0 +1,6 @@
+mod hyperdrive;
+mod network;
+mod schema;
+
+pub use hyperdrive::{in_memmory, Hyperdrive};
+pub use network::replicate_hyperdrive;
