+++ colmeia-hyperstack/Cargo.toml
@@ -1,16 +1,25 @@
 [package]
-name = 'colmeia-hypercore'
+name = 'colmeia-hyperstack'
 version = '0.1.0'
 authors = ['Bruno Tavares <connect+github@bltavares.com>']
 edition = '2018'
 
 [dependencies]
 futures = '0.3.5'
-hypercore = '0.11.1-beta.9'
+ed25519-dalek = '1.0.0-pre.3'
 anyhow = '1.0.34'
 random-access-storage = '4.0.0'
+random-access-memory = '2.0.0'
+random-access-disk = '2.0.0'
 log = '0.4.8'
-bitfield-rle = '0.2.0'
+hex = '0.4.2'
+thiserror = '1.0.20'
+
+[dependencies.colmeia-hyperswarm-mdns]
+path = '../colmeia-hyperswarm-mdns'
+
+[dependencies.colmeia-hyperdrive]
+path = '../colmeia-hyperdrive'
 
 [dependencies.async-std]
 version = '1.6.2'
