+++ colmeia-hyperstack/README.md
@@ -1,0 +1,7 @@
+# colmeia-hyperstack
+
+The whole stack that combines [datrs/hypercore](https://github.com/datrs/hypercore), [colmeia-hypercore](../colmeia-hypercore), [colmeia-hyperdrive](../colmeia-hyperdrive) and their discovery mechanisms into a single library to be ease to use on libs, CLIs and services.
+
+## Discovery mechanisms integrated
+
+- [colmeia-hyperswarm-mdns](../colmeia-hyperswarm-mdns)
