+++ colmeia-hyperstack/src/utils.rs
