+++ colmeia-hyperswarm-mdns/README.md
@@ -1,0 +1,3 @@
+# colmeia-hyperswarm-mdns
+
+Support to hyperswarm mdns infrastructure: based on [hyperswarm/discovery](https://github.com/hyperswarm/discovery/).
