+++ colmeiad/Cargo.toml
@@ -6,10 +6,8 @@ edition = '2018'
 
 [dependencies]
 tide = '0.12.0'
-log = '0.4.11'
 env_logger = '0.7.1'
 random-access-storage = '4.0.0'
-anyhow = '1.0.31'
 hypercore = '0.11.1-beta.9'
 futures = '0.3.5'
 
@@ -17,8 +15,8 @@ futures = '0.3.5'
 version = '1.0'
 features = ['derive']
 
-[dependencies.colmeia-hypercore]
-path = '../colmeia-hypercore'
+[dependencies.colmeia-hyperstack]
+path = '../colmeia-hyperstack'
 
 [dependencies.async-std]
 version = '1.6.0'
